Move selection blueprint to correct namespace
